Fix the UnsatisfiableMaxContains canonicalizer rule
