+++ src/canonicalizer/CMakeLists.txt
@@ -15,7 +15,8 @@ add_library(sourcemeta_jsonbinpack_canonicalizer
   rules/max_contains_without_contains.h
   rules/min_contains_without_contains.h
   rules/null_as_const.h
-  rules/then_else_without_if.h)
+  rules/then_else_without_if.h
+  rules/unsatisfiable_max_contains.h)
 target_include_directories(sourcemeta_jsonbinpack_canonicalizer PUBLIC
   "${CMAKE_CURRENT_SOURCE_DIR}/include")
 target_link_libraries(sourcemeta_jsonbinpack_canonicalizer PUBLIC sourcemeta_jsontoolkit_json)
