+++ src/canonicalizer/canonicalizer.cc
@@ -23,6 +23,7 @@
 #include "rules/min_contains_without_contains.h"
 #include "rules/null_as_const.h"
 #include "rules/then_else_without_if.h"
+#include "rules/unsatisfiable_max_contains.h"
 
 sourcemeta::jsonbinpack::Canonicalizer::Canonicalizer(
     const sourcemeta::jsontoolkit::schema_resolver_t &resolver)
@@ -45,6 +46,7 @@ sourcemeta::jsonbinpack::Canonicalizer::Canonicalizer(
   this->bundle.template add<MinContainsWithoutContains>();
   this->bundle.template add<NullAsConst>();
   this->bundle.template add<ThenElseWithoutIf>();
+  this->bundle.template add<UnsatisfiableMaxContains>();
 }
 
 auto sourcemeta::jsonbinpack::Canonicalizer::apply(
